Added auth for login and logout and register.

+++ src/client/js/controllers.js
@@ -1,8 +1,14 @@
-window.app.controller('PostController', ['$scope', 'postDataService', function($scope, postDataService) {
+window.app.controller('PostController', ['$scope', 'authService', 'postDataService', function($scope, authService, postDataService) {
   $scope.post = {};
   $scope.comment = {};
   $scope.showSubmit = false;
   $scope.showComments = false;
+  var user = authService.getUserInfo();
+  if (user) {
+    $scope.loggedIn = true;
+  } else {
+    $scope.loggedIn = false;
+  }
   $scope.addPost = function() {
     postDataService.addPost({
       title: this.post.title,
@@ -22,25 +28,29 @@ window.app.controller('PostController', ['$scope', 'postDataService', function($
   $scope.updatePosts = function() {
     postDataService.getAllPosts()
       .then(function(posts) {
-        console.log(posts);
         $scope.posts = posts.data.data;
       });
   };
   $scope.deletePost = function(id) {
-    postDataService.deletePost(id)
-      .then(function(post) {
-        $scope.updatePosts();
-      });
+    var ans = confirm('Are you sure that you want to delete this post?');
+    if (ans) {
+      postDataService.deletePost(id)
+        .then(function(post) {
+          $scope.updatePosts();
+        });
+    }
   };
   $scope.editPost = function(id) {
-    postDataService.editPost(id, this.post)
-      .then(function(post) {
-        $scope.updatePosts();
-      });
+    postDataService.getOnePost(id).then(function(post) {
+      $scope.post = post.data.data;
+      console.log($scope.post);
+      postDataService.editPost(id, $scope.post)
+        .then(function(post) {
+          $scope.updatePosts();
+        });
+    });
   };
   $scope.addComment = function(id) {
-    console.log('ID', id);
-    console.log('COMMENT_FORM:', this.comment);
     postDataService.addComment(id, {
       author: this.comment.name,
       message: this.comment.text,
@@ -48,13 +58,57 @@ window.app.controller('PostController', ['$scope', 'postDataService', function($
       .then(function(post) {
         $scope.updatePosts();
         $scope.comment = {};
+        post.showComments = true;
       });
   };
+  $scope.vote = function(id, direction) {
+    console.log('ID: ', id, 'direction: ', direction);
+    postDataService.getOnePost(id).then(function(post) {
+      console.log(post);
+      // if (direction === 'up') {
+      //   post.data.data.votes++;
+      // } else {
+      //   post.data.data.votes--;
+      // }
+      // postDataService.editPost(id, post.data.data).then(function(post) {
+      //   $scope.updatePosts();
+      // });
+    });
+  };
+
   $scope.updatePosts();
   $scope.sort = 'rating';
 }]);
-//
-// window.app.controller('mainController', function() {
-//   var vm = this;
-//   vm.time = new Date();
-// });
+
+app.controller('signupController', ['$scope', '$location', 'authService', function($scope,  $location, authService) {
+  $scope.user = {};
+  $scope.register = function() {
+      console.log($scope.user);
+      authService.register($scope.user)
+        .then(function(user) {
+          console.log('User: ', user);
+          authService.setUserInfo(user);
+          $location.path('/');
+        })
+        .catch(function(err) {
+          // TODO - check status code and show appropriate message.
+          console.log(err);
+        });
+  };
+}]);
+
+app.controller('loginController', ['$scope', '$location', '$window',  'authService', function($scope, $location, $window, authService) {
+  $scope.user = {};
+  $scope.login = function() {
+      authService.login($scope.user)
+        .then(function(user) {
+          console.log(user);
+          authService.setUserInfo(user);
+          $window.location.href = '/';
+        })
+        .catch(function(err) {
+          // TODO - check status code and show appropriate message.
+          console.log(err);
+        });
+  };
+}]);
